feat: fetchCoreEvalRelease
