+++ packages/boot/test/fast-usdc/fast-usdc.test.ts
@@ -12,13 +12,13 @@ import { unmarshalFromVstorage } from '@agoric/internal/src/marshal.js';
 import { defaultSerializer } from '@agoric/internal/src/storage-test-utils.js';
 import { eventLoopIteration } from '@agoric/internal/src/testing-utils.js';
 import { buildVTransferEvent } from '@agoric/orchestration/tools/ibc-mocks.js';
-import type { EndoZipBase64Bundle } from '@agoric/swingset-vat';
 import { makeRatio } from '@agoric/zoe/src/contractSupport/ratio.js';
 import { Fail } from '@endo/errors';
 import { makeMarshal } from '@endo/marshal';
 import NodeFetchCache, { FileSystemCache } from 'node-fetch-cache';
 import {
   AckBehavior,
+  fetchCoreEvalRelease,
   insistManagerType,
   makeSwingsetHarness,
 } from '../../tools/supports.js';
@@ -73,37 +73,6 @@ test.serial('oracles provision before contract deployment', async t => {
   t.truthy(watcherWallet);
 });
 
-const downloadCoreEval = async (
-  { fetch }: { fetch: typeof globalThis.fetch },
-  config = {
-    repo: 'Agoric/agoric-sdk',
-    release: 'fast-usdc-beta-1',
-    name: 'start-fast-usdc',
-  },
-  artifacts = `https://github.com/${config.repo}/releases/download/${config.release}`,
-  planUrl = `${artifacts}/${config.name}-plan.json`,
-) => {
-  const plan = (await fetch(planUrl).then(r => r.json())) as {
-    name: string;
-    permit: string;
-    script: string;
-    bundles: Array<{
-      bundleID: string;
-      entrypoint: string;
-      fileName: string;
-    }>;
-  };
-  assert.equal(plan.name, config.name);
-  const script = await fetch(`${artifacts}/${plan.script}`).then(r => r.text());
-  const permit = await fetch(`${artifacts}/${plan.permit}`).then(r => r.text());
-  const bundles: EndoZipBase64Bundle[] = await Promise.all(
-    plan.bundles.map(b =>
-      fetch(`${artifacts}/${b.bundleID}.json`).then(r => r.json()),
-    ),
-  );
-  return { bundles, evals: [{ js_code: script, json_permits: permit }] };
-};
-
 /**
  * Start with the Core Eval from proposal 87, the Fast USDC Beta release.
  *
@@ -115,9 +84,16 @@ const downloadCoreEval = async (
 test.serial('prop 87: Beta', async t => {
   const { evalProposal, bridgeUtils } = t.context;
 
-  const materials = await downloadCoreEval({
-    fetch: fetchCached,
-  });
+  const materials = await fetchCoreEvalRelease(
+    {
+      fetch: fetchCached,
+    },
+    {
+      repo: 'Agoric/agoric-sdk',
+      release: 'fast-usdc-beta-1',
+      name: 'start-fast-usdc',
+    },
+  );
 
   // Proposal 87 doesn't quite complete: noble ICA is mis-configured
   bridgeUtils.setAckBehavior(
